pyproject.toml
[tool.poetry]
name = "saleae_resampler"
version = "0.1.0"
description = "My \"brilliant\" desc"
authors = ["Joni Kähärä <joni.kahara@async.fi>"]
homepage = "https://github.com/kahara/python-saleae-resampler/"
repository = "https://github.com/kahara/python-saleae-resampler/"
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
saleae_resampler = "saleae_resampler.console:saleae_resampler_cli"

[tool.black]
line-length = 120
target-version = ['py37']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  |  __pycache__
)
'''

[tool.pytest.ini_options]
junit_family="xunit2"
addopts="--cov=saleae_resampler --cov-fail-under=65 --cov-branch"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-parents = 10

[tool.pylint.messages_control]
disable=["fixme", "W1202", "C0330"]

[tool.coverage.run]
omit = ["tests/*"]
branch = true

[tool.poetry.dependencies]
python = "^3.8"

python-dateutil = "^2.8.2"
pyaml = "^21.8.3"
libadvian = "^0.2.0"
[tool.poetry.dev-dependencies]
pytest = "^6.2"
coverage = {version = "^5.4", extras = ["toml"]}
pytest-cov = "^2.11"
pylint = "^2.5"
black = "=20.8b1"
bandit = "^1.6"
mypy = "^0.800"
pre-commit = "^1.20"
pytest-asyncio = "^0.14"
bump2version = "^1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
